Fully fix kick bug for all versions.
The problem was that if a scoreboard team is sent while it's already registered, the player will disconnect.
Since there does not appear to be any downside to never removing the team,
we just won't remove the team, and send the team only once.
If there are any problems with this in the future, we can implement a smart solution which will be able to handle the packet sending properly.